added error page display for certain actions

+++ Hearts of Gold/Hearts of Gold/Controllers/ItemsController.cs
@@ -134,7 +134,8 @@ namespace Hearts_of_Gold.Controllers
 
             if (userId != userCompareId || itemId != item.ItemID)
             {
-                return RedirectToAction("Details", new { id = item.ItemID });
+                ViewData["error"] = "You can only edit your own item";
+                return View("Error");
             }
 
             if (ModelState.IsValid)
@@ -153,6 +154,7 @@ namespace Hearts_of_Gold.Controllers
         // GET: Items/Delete/5
         public ActionResult Delete(int? id)
         {
+
             if (id == null)
             {
                 return new HttpStatusCodeResult(HttpStatusCode.BadRequest);
@@ -171,6 +173,15 @@ namespace Hearts_of_Gold.Controllers
         public ActionResult DeleteConfirmed(int id)
         {
             Item item = db.Items.Find(id);
+            var userId = ReturnUserId();
+            var userCompareId = db.Items.Where(i => i.ItemID == item.ItemID).Select(i => i.UserID).FirstOrDefault();
+            var itemId = db.Items.Where(i => i.ItemID == item.ItemID).Select(i => i.ItemID).FirstOrDefault();
+
+            if (userId != userCompareId || itemId != item.ItemID)
+            {
+                ViewData["error"] = "You can only delete your own item";
+                return View("Error");
+            }
             db.Items.Remove(item);
             db.SaveChanges();
             return RedirectToAction("Index");
